+++ CHANGELOG.md
@@ -1,7 +1,21 @@
 
-# master
+## [v2.0](https://github.com/c-sp/gameboy-test-roms/releases/tag/v2.0) *(2020-07-09)*
 
-# v1.1
+- use
+  [dmg-acid2:8a98ce731f](https://github.com/mattcurrie/dmg-acid2/tree/8a98ce731f)
+  *(2020-04-21)*
+- use
+  [cgb-acid2:76c4e2c42d](https://github.com/mattcurrie/cgb-acid2/tree/76c4e2c42d)
+  *(2020-04-21)*
+- use
+  [mealybug-tearoom-tests:844b92ea59](https://github.com/mattcurrie/mealybug-tearoom-tests/tree/844b92ea59)
+  *(2019-10-24)*
+- compile dmg-acid2, cgb-acid2 and mealybug-tearoom-tests with
+  [RGBDS:8b60efa149](https://github.com/rednex/rgbds/tree/8b60efa149)
+  *(2020-06-27)*
+
+
+## [v1.1](https://github.com/c-sp/gameboy-test-roms/releases/tag/v1.1) *(2020-06-14)*
 
 - use
   [mooneye-gb:6b9488fa3e](https://github.com/Gekkio/mooneye-gb/tree/6b9488fa3e)
@@ -15,7 +29,8 @@
   [gambatte:56e3371151](https://github.com/sinamas/gambatte/tree/56e3371151)
   *(2020-03-08)*
 
-# v1.0
+
+## [v1.0](https://github.com/c-sp/gameboy-test-roms/releases/tag/v1.0) *(2020-06-13)*
 
 - compile Mooneye GB test roms with
   [wla-dx:d9a3237cc2](https://github.com/vhelin/wla-dx/tree/d9a3237cc2)